changed to fp32

+++ scripts/train_classification.py
@@ -246,7 +246,7 @@ def training_state(model_path, model, rank):
 
 
 def main():
-    torch.set_default_dtype(torch.float16)
+    torch.set_default_dtype(torch.float32)
 
     print0("Loading model...")
     model = models.get_model(
